sentimental_model/requirements.txt
absl-py==1.4.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==5.3.0
certifi==2020.12.5
chardet==3.0.4
click==8.1.2
codeswitch==1.1
colorama==0.4.4
cycler==0.11.0
Django==3.1.7
django-templatetag-randomgen==0.3
Faker==8.0.0
filelock==3.9.0
Flask==2.1.1
Flask-Cors==3.0.10
flatbuffers==23.1.21
fonttools==4.32.0
gast==0.4.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
googletrans==3.1.0a0
graphviz==0.20.1
greenlet==1.1.2
grpcio==1.51.1
gspread==3.7.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.19.1
httpx==0.13.3
huggingface-hub==0.12.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.0.1
jinja2-django-tags==0.5
joblib==1.1.0
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==15.0.6.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
nltk==3.8.1
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
protobuf==3.19.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pyparsing==2.4.7
pyspellchecker==0.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
regex==2022.10.31
requests==2.25.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.15.0
sklearn==0.0
sniffio==1.3.0
SQLAlchemy==1.4.36
sqlparse==0.4.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0